requirements.txt
cycler==0.11.0
fonttools==4.34.4
glcontext==2.3.6
GPUtil==1.4.0
kiwisolver==1.4.4
matplotlib==3.5.2
moderngl==5.6.4
moderngl-window==2.4.1
MouseInfo==0.1.3
multipledispatch==0.6.0
packaging==21.3
Pillow==9.3.0
psutil==5.9.1
py-cpuinfo==8.0.0
PyAutoGUI==0.9.53
pygame==2.1.2
PyGetWindow==0.0.9
pyglet==1.5.26
pyjoystick==1.2.4
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrr==0.10.3
PyScreeze==0.1.28
PySDL2==0.9.12
python-dateutil==2.8.2
pytweening==1.0.4
PyWavefront==1.3.3
resource-man==2.2.8
six==1.16.0
